New url function integrated on common module
